Refactor classes so that functionality can be reused by monitoring  extraction.
Split out configuration and input classes.
Split out service functionality.
Add formatter for JSON
Fix old SLF4J logging declaration
Clean up some unused imports

Additional commit.
