+++ src/main/java/com/elastic/support/diagnostics/DiagConfig.java
@@ -1,26 +1,22 @@
-package com.elastic.support.config;
+package com.elastic.support.diagnostics;
 
-import com.elastic.support.util.JsonYamlUtils;
+import com.elastic.support.config.BaseConfig;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.util.List;
 import java.util.Map;
 
-public class DiagConfig {
+public class DiagConfig extends BaseConfig {
 
     private static Logger logger = LogManager.getLogger(DiagConfig.class);
 
 
-    private Map<String, String> githubSettings;
     private Map<String, Integer> logSettings;
-    private List<String> passwordKeys;
     private List<String> textFileExtensions;
-    private Map<String, Integer> restConfig;
     private int callRetries, pauseRetries;
     private List<String> requireRetry;
     private Map restCalls;
-    private Map configuration;
 
 
     public DiagConfig(){
@@ -29,26 +25,26 @@ public class DiagConfig {
 
     public DiagConfig(Map configuration) {
 
-        this.configuration = configuration;
+        super(configuration);
 
-        githubSettings = (Map<String, String>) configuration.get("github-settings");
+        // Settings for REST calls to check for new diagnostic version
+
+        // How many rolled over logs do we get?
         logSettings = (Map<String, Integer>) configuration.get("log-settings");
-        passwordKeys = (List<String>) configuration.get("password-keys");
+
+        // Differentiate between json and plain text results
         textFileExtensions = (List<String>) configuration.get("text-file-extensions");
-        restConfig = (Map<String, Integer>) configuration.get("rest-config");
+
+        // REST settings for HttpClient
+
+        // Whether we retry a failed call, how many times, and how long to wait before reattempting.
+        requireRetry = (List<String>)configuration.get("require-retry");
         callRetries = (Integer) configuration.get("call-retries");
         pauseRetries = (Integer)configuration.get("pause-retries");
-        requireRetry = (List<String>)configuration.get("require-retry");
+
+        // The REST calls we execute
         restCalls = (Map) configuration.get("rest-calls");
 
-    }
-
-    public Map<String, String> getGithubSettings() {
-        return githubSettings;
-    }
-
-    public void setGithubSettings(Map<String, String> githubSettings) {
-        this.githubSettings = githubSettings;
     }
 
     public Map<String, Integer> getLogSettings() {
@@ -59,28 +55,12 @@ public class DiagConfig {
         this.logSettings = logSettings;
     }
 
-    public List<String> getPasswordKeys() {
-        return passwordKeys;
-    }
-
-    public void setPasswordKeys(List<String> passwordKeys) {
-        this.passwordKeys = passwordKeys;
-    }
-
     public List<String> getTextFileExtensions() {
         return textFileExtensions;
     }
 
     public void setTextFileExtensions(List<String> textFileExtensions) {
         this.textFileExtensions = textFileExtensions;
-    }
-
-    public Map<String, Integer> getRestConfig() {
-        return restConfig;
-    }
-
-    public void setRestConfig(Map<String, Integer> restConfig) {
-        this.restConfig = restConfig;
     }
 
     public int getCallRetries() {
@@ -113,14 +93,6 @@ public class DiagConfig {
 
     public void setRestCalls(Map restCalls) {
         this.restCalls = restCalls;
-    }
-
-    public Map getConfiguration() {
-        return configuration;
-    }
-
-    public void setConfiguration(Map configuration) {
-        this.configuration = configuration;
     }
 
     public Map<String, String> getCommandMap(String key){
