+++ src/main/java/com/elastic/support/diagnostics/DiagnosticService.java
@@ -1,9 +1,8 @@
 package com.elastic.support.diagnostics;
 
 import com.elastic.support.BaseService;
+import com.elastic.support.ElasticClientService;
 import com.elastic.support.config.Constants;
-import com.elastic.support.config.DiagConfig;
-import com.elastic.support.config.DiagnosticInputs;
 import com.elastic.support.diagnostics.chain.DiagnosticChainExec;
 import com.elastic.support.diagnostics.chain.DiagnosticContext;
 import com.elastic.support.rest.RestClient;
@@ -14,13 +13,12 @@ import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang3.StringUtils;
 
 import java.io.File;
-import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.List;
 import java.util.Map;
 
-public class DiagnosticService extends BaseService {
+public class DiagnosticService extends ElasticClientService {
 
 
     public void exec(DiagnosticInputs diagnosticInputs, DiagConfig diagConfig, Map<String, List<String>> chains) {
@@ -33,13 +31,21 @@ public class DiagnosticService extends BaseService {
         try {
 
             // Create the temp directory - delete if first if it exists from a previous run
-            String tempDir = diagnosticInputs.getTempDir();
+            String outputDir = diagnosticInputs.getOutputDir();
+            String tempDir;
+            if (diagnosticInputs.getDiagType().equals(Constants.ES_DIAG_DEFAULT)) {
+                tempDir = outputDir + SystemProperties.fileSeparator + Constants.ES_DIAG;
+
+            } else {
+                tempDir = outputDir + SystemProperties.fileSeparator + diagnosticInputs.getDiagType() + "-" + Constants.ES_DIAG;
+            }
+
             logger.info("Creating temp directory: {}", tempDir);
 
             FileUtils.deleteDirectory(new File(tempDir));
             Files.createDirectories(Paths.get(tempDir));
 
-            logger.info("Created temp directory: {}", diagnosticInputs.getTempDir());
+            logger.info("Created temp directory: {}", tempDir);
 
             // Set up the log file manually since we're going to package it with the diagnostic.
             // It will go to wherever we have the temp dir set up.
@@ -50,39 +56,24 @@ public class DiagnosticService extends BaseService {
             List<String> chain = chains.get(diagnosticInputs.getDiagType());
 
             DiagnosticChainExec dc = new DiagnosticChainExec();
-            ctx.setEsRestClient(esRestClient);
-            ctx.setGenericClient(genericClient);
+            ctx.setEsRestClient(createEsRestClient(diagConfig, diagnosticInputs));
+            ctx.setGenericClient(createGenericClient(diagConfig, diagnosticInputs));
             ctx.setDiagsConfig(diagConfig);
             ctx.setDiagnosticInputs(diagnosticInputs);
             ctx.setTempDir(tempDir);
 
             dc.runDiagnostic(ctx, chain);
 
-            if (ctx.isDocker()){
+            if (ctx.isDocker()) {
                 logger.warn("Identified Docker installations - bypassed log collection and system calls.");
             }
 
-            String user = ctx.getDiagnosticInputs().getUser();
+           checkAuthLevel(ctx.getDiagnosticInputs().getUser(), ctx.isAuthorized());
 
-            if (StringUtils.isNotEmpty(user) && !ctx.isAuthorized()) {
-
-                String border = SystemUtils.buildStringFromChar(60, '*');
-                logger.warn(SystemProperties.lineSeparator);
-                logger.warn(border);
-                logger.warn(border);
-                logger.warn(border);
-                logger.warn("The elasticsearch user entered: {} does not appear to have sufficient authorization to access all collected information", user);
-                logger.warn("Some of the calls may not have completed successfully.");
-                logger.warn("If you are using a custom role please verify that it has the admin role for versions prior to 5.x or the superuser role for subsequent versions.");
-                logger.warn(border);
-                logger.warn(border);
-                logger.warn(border);
-
-            }
         } catch (DiagnosticException de) {
             logger.warn(de.getMessage());
         } catch (Throwable t) {
-            logger.log(SystemProperties.DIAG, "Temp directory error" , t);
+            logger.log(SystemProperties.DIAG, "Temp directory error", t);
             logger.warn(String.format("Issue with creating temp directory. %s", Constants.CHECK_LOG));
         } finally {
             closeLogs();
@@ -93,48 +84,8 @@ public class DiagnosticService extends BaseService {
         }
     }
 
-    private RestClient createGenericClient(DiagConfig diagConfig, DiagnosticInputs diagnosticInputs) {
-        RestClientBuilder builder = new RestClientBuilder();
 
-        return builder
-                .setConnectTimeout(diagConfig.getRestConfig().get("connectTimeout") * 1000)
-                .setRequestTimeout(diagConfig.getRestConfig().get("requestTimeout") * 1000)
-                .setSocketTimeout(diagConfig.getRestConfig().get("socketTimeout") * 1000)
-                .setProxyHost(diagnosticInputs.getProxyUser())
-                .setProxPort(diagnosticInputs.getProxyPort())
-                .setProxyUser(diagnosticInputs.getUser())
-                .setProxyPass(diagnosticInputs.getProxyPassword())
-                .build();
 
-    }
-
-    private RestClient createEsRestClient(DiagConfig diagConfig, DiagnosticInputs diagnosticInputs) {
-        RestClientBuilder builder = new RestClientBuilder();
-        builder
-                .setConnectTimeout(diagConfig.getRestConfig().get("connectTimeout") * 1000)
-                .setRequestTimeout(diagConfig.getRestConfig().get("requestTimeout") * 1000)
-                .setSocketTimeout(diagConfig.getRestConfig().get("socketTimeout") * 1000)
-                .setProxyHost(diagnosticInputs.getProxyUser())
-                .setProxPort(diagnosticInputs.getProxyPort())
-                .setProxyUser(diagnosticInputs.getUser())
-                .setProxyPass(diagnosticInputs.getProxyPassword())
-                .setBypassVerify(diagnosticInputs.isSkipVerification())
-                .setHost(diagnosticInputs.getHost())
-                .setPort(diagnosticInputs.getPort())
-                .setScheme(diagnosticInputs.getScheme());
-
-        if (diagnosticInputs.isSecured()) {
-            builder.setUser(diagnosticInputs.getUser())
-                    .setPassword(diagnosticInputs.getPassword());
-        }
-
-        if (diagnosticInputs.isPki()) {
-            builder.setPkiKeystore(diagnosticInputs.getPkiKeystore())
-                    .setPkiKeystorePass(diagnosticInputs.getPkiKeystorePass());
-        }
-
-        return builder.build();
-    }
 
 
 }
