+++ src/main/java/com/elastic/support/diagnostics/commands/BaseQueryCmd.java
@@ -1,6 +1,6 @@
 package com.elastic.support.diagnostics.commands;
 
-import com.elastic.support.config.DiagConfig;
+import com.elastic.support.diagnostics.DiagConfig;
 import com.elastic.support.diagnostics.chain.Command;
 import com.elastic.support.rest.RestCallManifest;
 import com.elastic.support.rest.RestClient;
