+++ src/main/java/com/elastic/support/diagnostics/commands/VersionCheckCmd.java
@@ -1,7 +1,6 @@
 package com.elastic.support.diagnostics.commands;
 
 import com.elastic.support.config.Constants;
-import com.elastic.support.config.DiagnosticInputs;
 import com.elastic.support.config.Version;
 import com.elastic.support.diagnostics.DiagnosticException;
 import com.elastic.support.diagnostics.chain.Command;
