+++ src/main/java/com/elastic/support/scrub/ScrubInputs.java
@@ -1,6 +1,7 @@
-package com.elastic.support.config;
+package com.elastic.support.scrub;
 
 import com.beust.jcommander.Parameter;
+import com.elastic.support.config.BaseInputs;
 import com.elastic.support.util.SystemProperties;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.LogManager;
