+++ src/main/java/com/elastic/support/scrub/ScrubService.java
@@ -1,9 +1,7 @@
 package com.elastic.support.scrub;
 import com.elastic.support.config.Constants;
-import com.elastic.support.config.ScrubInputs;
 import com.elastic.support.BaseService;
 import com.elastic.support.util.ArchiveUtils;
-import com.elastic.support.util.JsonYamlUtils;
 import com.elastic.support.util.SystemProperties;
 import com.elastic.support.util.SystemUtils;
 import org.apache.commons.io.FileUtils;
